[tests] Remove spurious error log in p2p_segwit.py
Since b5e4ea5a0227cbed8ce1a07b7ff33e0f33f35c61, when wait_until() fails,
an error message is logged to the test framework log. This means that if
wait_until() is called inside a try-except with the expectation that it
will fail, a spurious error message is logged.

wait_until() shouldn't be called with the expectation of failure. Fix
that in p2p_segwit.py.