Removed explicit mention of storage requirement (squashed)
Similarly discussed and fixed in the following bitcoin.org issue: https://github.com/bitcoin-dot-org/bitcoin.org/pull/2716

Removed mention of storage requirement

Agreeing with @laanwj that this mention of the storage requirement could be removed, I did so and changed the wording accordingly.
Would be nice to be able to add a link to a canonical source that is always updated, for those who want a number.

Update doc/README.md

Co-Authored-By: merland <martin@megabit.se>

Update README.md

added a missing 'a'