[addrman] Fix new table bucketing during unserialization
An addrman entry can appear in up to 8 new table buckets. We store this
entry->bucket indexing during shutdown so that on restart we can restore
the entries to their correct buckets.

Commit 409c188dfee14cbb5a858481e2130da93411e193 broke the
deserialization code so that each entry could only be put in up to one
new bucket. Fix that.