Make CAddrman::Select_ select buckets, not positions, first
The original CAddrMan behaviour (before commit
40a09c12891d79d819f5be67198a561e7bef631b) was to pick a uniformly
random non-empty bucket, and then pick a random element from that
bucket. That commit, which introduced deterministic placement
of entries in buckets, changed this to picking a uniformly random
non-empty bucket position instead.

This commit reverts the original high-level behavior, but in the
deterministic placement model.