test: add regression tests for #27468 (invalid URI segfaults)
Prior to PR #27468 (commit 4bd72fe04140175089107b7e76d6822607136bad) all
call-sites of `GetQueryParameter(...)` in the REST module could trigger
a crash. Add missing test cases for all possible code-paths as a
regression test, as a foundation for possible follow-up fixes (which aim
to resolve this issue in a more general and robust way).